invoice indexpage updates

+++ Models/Invoice.cs
@@ -10,29 +10,29 @@ namespace Alliance_for_Life.Models
         public string OrgName { get; set; }
 
         [UIHint("Currency")]
-        public decimal DirectAdminCost { get; set; }
+        public double DirectAdminCost { get; set; }
 
         [UIHint("Currency")]
-        public decimal ParticipantServices { get; set; }
+        public double ParticipantServices { get; set; }
 
         [UIHint("Currency")]
-        public decimal GrandTotal { get; set; }
+        public double GrandTotal { get; set; }
 
         public double LessManagementFee { get; set; }
 
         [UIHint("Currency")]
-        public decimal DepositAmount { get; set; }
+        public double DepositAmount { get; set; }
 
         [UIHint("Currency")]
-        public decimal BeginningAllocation { get; set; }
+        public double BeginningAllocation { get; set; }
 
         [UIHint("Currency")]
-        public decimal AdjustedAllocation { get; set; }
+        public double AdjustedAllocation { get; set; }
 
         public DateTime BillingDate { get; set; }
 
         [UIHint("Currency")]
-        public decimal BalanceRemaining { get; set; }
+        public double BalanceRemaining { get; set; }
 
         public DateTime SubmittedDate { get; set; }
